Introduce new continueOnAssertionFailure maven plugin option
Motivation:

We created test set with goal to run it on weekend, or during night.
We expected to get all results and make conclusions based on this results, but blocked issue was appeared, assertion failure breaks all simulations.
So we decided to make changes in compliance with our goals.

Modifications:

Was added new property - "continueOnAssertionFailure" with default valued "false".
This property checked during  handling of "GatlingSimulationAssertionsFailedException" and if it was set to true, save exception and continue simulations.
In the very end we see warning in console - "There were some errors while running your simulation, but continueOnAssertionFailure was set to true, so your simulations continue to perform.", but all reports saves and we can get all information we need.

If property was set to "false" - previous (base) logic is preserved, simulations stops.
Also was changed Collection interface by List, cause it more preferable to work with List (you don't need to transform it in array if you want to get object by index).

Result:

Goal was reached, if property was set to "true" we get all reports and can make conclusions based on it.