Improve default simulationsFolder for the Recorder
Motivation:
When running the recorder from the plugin, the simulationsFolder currently always defaults to the standard Java test sources root (even in a project set up for Kotlin or Scala).

Modifications:
Use the first element in project.testCompileSourceRoots as the default simulationsFolder.

Result:
The recorder will default to saving simulations to the folder set with <testSourceDirectory> in the POM, which is typically what the user would expect.