feat: switch interactive mode based on maven --batch-mode option or CI env var
Motivation:

The interactive mode of the old bundle was very helpful.
We should replicate it in the maven plugin now that it's replacing the old bundle.

Modification:

Unless maven batch mode is enable or the CI env var is set to true, switch to interactive mode and let the user select the simulation class and the run description.

Interactive mode will be completely dropped from core Gatling.